feature_qrs_experiment/requirements.txt
wfdb>=4.1.2
numpy>=1.22
scikit-learn>=1.3
scipy>=1.10
tqdm>=4.66
joblib>=1.3
pandas>=2.0
torch>=2.1
pyyaml>=6.0
